- Fixed Menu_Bar redraw bug.  - Changed Fl_Choice to work way it works in fltk2  - Fixed many menu related minor bugs  - Fixed all boxtype names  - Added couple of new boxtypes to EFluid

+++ src/core/fl_glyph.cpp
@@ -72,7 +72,8 @@ int x,int y,int w,int h, Fl_Flags flags)
     {
         fl_color(i ? Fl_Color(FL_LIGHT3) : color);
 
-        int w1 = (w+2)/3; int x1,y1;
+        int w1 = (w+2)/3;
+        double x1,y1;
         switch(glyph)
         {
